Avkr003/validate (#963)
* change discount values

* more changes

* get discount in amm:

* ifx tests

* fix tests

* fix

* tradeshield fix

* use tier structure

* update tests

* fixing tier keeper pointers (#958)

* fixing tier keeper pointers

* fixing tier keeper pointers

* fixing tier keeper pointers

* fix

* fix

* add discoumt to perpetual

* fix tests

* review changes

* use silver

* adding validation for messages and params

* merging discount branch

* fixing test cases

* fixing test cases

* reverting back vesting tx cmd

* fixing test case

* increasing test coverage

* adding test cases

* Enforce stricter validation on TakeProfitPriceRatios

Added checks to ensure MaximumLongTakeProfitPriceRatio is greater than 1 and greater than MinimumLongTakeProfitPriceRatio, while MinimumLongTakeProfitPriceRatio must also be greater than 1. Additionally, imposed that MaximumShortTakeProfitPriceRatio must be less than 1 to maintain consistency.

* Add validation checks for parameter

Ensure that higher-bound parameters are greater than lower-bound counterparts in stablestake and perpetual modules. Also, add a combined reward portion check in the masterchef module and improve clarity in error messages for the oracle module.

---------

+++ x/masterchef/keeper/abci.go
@@ -45,7 +45,7 @@ func (k Keeper) GetPoolTVL(ctx sdk.Context, poolId uint64) math.LegacyDec {
 func (k Keeper) ProcessExternalRewardsDistribution(ctx sdk.Context) {
 	baseCurrency, _ := k.assetProfileKeeper.GetUsdcDenom(ctx)
 	curBlockHeight := ctx.BlockHeight()
-	totalBlocksPerYear := k.parameterKeeper.GetParams(ctx).TotalBlocksPerYear
+	totalBlocksPerYear := int64(k.parameterKeeper.GetParams(ctx).TotalBlocksPerYear)
 
 	externalIncentives := k.GetAllExternalIncentives(ctx)
 	externalIncentiveAprs := make(map[uint64]math.LegacyDec)
@@ -170,7 +170,7 @@ func (k Keeper) UpdateLPRewards(ctx sdk.Context) error {
 	totalProxyTVL, totalProxyTvlEdenEnable := k.CalculateProxyTVL(ctx, baseCurrency)
 
 	// Ensure totalBlocksPerYear is not zero to avoid division by zero
-	totalBlocksPerYear := k.parameterKeeper.GetParams(ctx).TotalBlocksPerYear
+	totalBlocksPerYear := int64(k.parameterKeeper.GetParams(ctx).TotalBlocksPerYear)
 	if totalBlocksPerYear == 0 {
 		return errorsmod.Wrap(types.ErrNoInflationaryParams, "invalid inflationary params")
 	}
@@ -267,7 +267,7 @@ func (k Keeper) UpdateLPRewards(ctx sdk.Context) error {
 		dataLifetime := params.RewardsDataLifetime
 		for {
 			firstAccum := k.FirstPoolRewardsAccum(ctx, pool.PoolId)
-			if firstAccum.Timestamp == 0 || int64(firstAccum.Timestamp)+dataLifetime >= ctx.BlockTime().Unix() {
+			if firstAccum.Timestamp == 0 || int64(firstAccum.Timestamp+dataLifetime) >= ctx.BlockTime().Unix() {
 				break
 			}
 			k.DeletePoolRewardsAccum(ctx, firstAccum)
